Introduce a severity level for issues, and a 'warning' severity (un-revert #931) (#952)
This un-reverts #950, effectively reintroducing the changes recently
landed in #931.

The revert was needed because it revealed a latent bug in the Swift
compiler, tracked by https://github.com/swiftlang/swift/issues/79304. I
reproduced that failure and included a workaround in the second commit
on this PR.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.