Cleaning recently changed docs (#924)
Improving the style and explanation of recent additions to the
documentation.

The recent changes to test traits are great, and come with some very
nice documentation. I've done what I can to improve the style and the
wording to make it easier to understand, and to make the style
consistent with other documentation in the project.

Changed the documentation for `Trait` and associated types.

There are no behavioral changes in this PR, but I hope that the existing
behavior is easier to use with better documentation.

- [X] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [X] If public symbols are renamed or modified, DocC references should
be updated.