Plumb the schema version through the JSON ABI layers. (#956)
This PR plumbs the schema version (0 or 1) through the Swift types and
functions that produce our JSON output, allowing them to change
behaviour based on which version is in use.

As a proof-of-concept as much as anything else, this PR also adds an
(unsupported) `_tags` field to the JSON output for a test record, but
only with ABI version 1.

I've split this PR into two commits: the first plumbs the version
through as an integer argument, while the second uses the type system to
represent different ABI versions. The latter uses a protocol that we can
extend so that different ABI versions have differing behaviours
(although that's too coarse-grained for things like `_tags`.)

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.