Avoid calling `unsafeBitCast(_:to:)` to cast C function pointers.
This PR replaces calls to `unsafeBitCast(_:to:)` (where the value being cast is
a C function pointer) with calls to an internal function that checks that the
value being cast is actually a C function pointer. This is intended to make the
code in question more self-documenting (by making it clear what kind of cast is
being performed.)

Also fixed one place we're using `unsafeBitCast(_:to:)` to cast a metatype
where, as of Swift 6.1, we don't need to anymore.