Rename `ABI.v1` to `ABI.v6_2`.
We just added `ABI.v1`, but to help developers understand which version of Swift
it will be available in, we're thinking we want to rename it to `v6_2`.

The integer representation of the version, which is generally only of interest
to tools authors, is `60200`. We could allow a string in this position but it
would require a change in SwiftPM as well as a new type in Swift Testing that
can be converted to/from both an integer and a string, and that level of
complexity is hard to justify given the ABI version number's relatively niche
use cases. We can revisit this decision in the future (before Swift 6.2 is
released) if we change our minds.