build: restructure the installation location
Clean up some of the module property handling to be more declarative.
This reduces the operations that are being done on targets in favour
of defaults.

Adopt the latest best practices for handling module installation.
Introduce the new `SwiftTesting_INSTALL_NESTED_SUBDIR` option to allow
installation into platform/architecture subdirectory allowing
multi-architecture installations for platforms like Windows and Android.
This is currently opt-in and requires a newer toolchain (something
within the last ~2w) to detect the defaults. The values can be
overridden by the user if desired.