Change the return types of test content accessors and add `ExitTest.current`.
This PR changes the signatures of the accessor functions for the test content
metadata section so that we're not using underscored types. Instead, a
`Test.Generator` type replaces `Test._Record` and the bare function type we were
using for `Test` and `ExitTest` is promoted to (still experimental) API instead
of hiding behind underscores.

This change allows the types produced by accessor functions to always be nominal
types, avoids the iffy substitution of `_Record` for `@Sendable () async -> Test`
at runtime, and simplifies the internal `TestContent` protocol by eliminating
the need for an associated type to stand in for the conforming type.

This change leaves `ExitTest` as a nominal, but empty, type. I've added
`ExitTest.current` to make it useful so that test code, especially libraries,
can detect if it's running inside an exit test or not. I will separately update
the exit test proposal PR to include this type and property.