Show the number of test cases that ran at the end of a test (#972)
### Motivation:

To have the number of test cases at the end of the test in for
parameterized tests, so the logs don’t easily get buried.

### Modifications:

Log the number of test cases for parameterized tests at the end of a
test.

### Result:

If the user runs the parameterized tests, an extra piece of information
(with * test cases) would get bound to the end log.

Resolves #943.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.