Rename `ExitTestArtifacts` and split `ExitCondition` in twain.
This PR renames `ExitTestArtifacts` to `ExitTest.Result` and splits
`ExitCondition` into two types: `ExitTest.Condition` which can be passed to
`#expect(exitsWith:)` and `StatusAtExit` which represents the raw, possibly
platform-specific status reported by the kernel when a child process terminates.

The latter type is not nested in `ExitTest` because it can be used independently
of exit tests and we may want to use it in the future for things like
multi-process parallelization, but if a platform supports spawning processes but
not exit tests, nesting it in `ExitTest` would make it unavailable.

I considered several names for `StatusAtExit`:

- `ExitStatus`: too easily confusable with exit _codes_ such as `EXIT_SUCCESS`;
- `ProcessStatus`: we don't say "process" in our API surface elsewhere;
- `Status`: too generic
- `ExitReason`: "status" is a more widely-used term of art for this concept.

Foundation uses `terminationStatus` to represent the raw integer value and
`Process.TerminationReason` to represent whether it's an exit code or signal. We
don't use "termination" in Swift Testing's API anywhere.

I settled on `StatusAtExit` because it was distinct and makes it clear that it
represents the status of a process _at exit time_ (as opposed to while running,
e.g. `enum ProcessStatus { case running; case suspended; case terminated }`.

Updates to the exit tests proposal document will follow in a separate PR.