Diagnose when using a non-escapable type as suite. (#988)
This PR introduces a custom compile-time diagnostic when attempting to
use a non-escapable type as a suite. For example:

```swift
@Suite struct NumberOfBeesTests: ~Escapable {
  @Test borrowing func countBees() { ... } // 🛑 Attribute 'Test' cannot be applied to a function within structure 'NumberOfBeesTests' because its conformance to 'Escapable' has been suppressed
}
```

Values with non-escapable type cannot currently be initialized nor
returned from a function, and we need to be able to do both in order to
correctly implement the `@Test` macro.

This change does not diagnose if `@Suite` is applied to such a type but
does not contain any test functions, because we do compile successfully
in that case and this sort of pattern remains valid:

```swift
@Suite struct MyTests: ~Escapable {
  @Suite struct EndToEndTests {
    ...
  }
}
```

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.