Refine type-based test discovery mechanism to use test content records.
This PR changes how we discover tests in the type metadata section to more
closely align with how the test content section works. This change will allow
for a smoother transition to the test content section and our use of `@section`
by using the same underlying structures (as much as is feasible.) Client code
that uses the new `_TestDiscovery` module will need fewer changes to adapt.