[CMake] Set CMP0157 to OLD when targeting Android with the Windows toolchain (#1009)
There is no early swift-driver build for the Windows toolchain. As a
result, swift-testing fails to build properly when CMP0157 is set to NEW
due to object files not being generated.

This sets CMP0157 to OLD when targeting Android with the Windows
toolchain until the early swift-driver is available on Windows. This is
analog to
https://github.com/swiftlang/swift-corelibs-foundation/pull/5180

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.