Fix crash when loading test content records from the type metadata section.
Looks like # caused a regression: the compiler appears to be dead-code-stripping
the classes we emit to contain test content records. This PR changes the design
back to using a protocol so that the members we need are always covered by
`@_alwaysEmitConformanceMetadata`. This makes `_TestDiscovery` a little harder
to use with legacy lookup, but it's all experimental and eventually going to be
removed anyway.

Resolves rdar://146809312.