Add a dedicated `TestContentKind` type to `_TestDiscovery`. (#1019)
This PR adds a `TestContentKind` type to `_TestDiscovery` that
represents the `kind` field of a record and allows initialization from a
string literal (i.e. a FourCC) or an integer literal:

```swift
extension MyType: DiscoverableAsTestContent {
  static var testContentKind: TestContentKind {
    "moof"
  }
}
```

At this time, it's not possible to use this type directly in test
content records emitted by our macros because the type cannot be made
public, but it allows for types conforming to
`DiscoverableAsTestContent` to be a bit more expressive. Test library
authors who are experimenting with `_TestDiscovery` may opt to export
the symbols from `_TestDiscovery` and thus could use the type directly
in their emitted records (but keep in mind this module is still
experimental!)

The new type is `@frozen` and mostly `@inlinable` because its layout is
set in stone by virtue of the test content record `kind` field having a
fixed layout.

Resolves rdar://146855125.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.