Continue encoding value for deprecated property in runner plan snapshot to avoid decoding error in clients (#1023)
This fixes a regression in the encoding of the `Runner.Plan` snapshot
types, which can manifest when using Xcode 16.

### Motivation:

The `isParallelizationEnabled` property was deprecated and changed from
a stored property to a derived one in #901. That caused the value to no
longer be encoded in `Runner.Plan.Action.RunOptions`, which can cause
decoding errors in versions of Xcode which expect it to still be
present.

### Modifications:

- Manually implement `Codable` conformance for the affected type to
begin including a hardcoded value.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.

Fixes rdar://146284519