Set CMP0157 to OLD for Android
There is no early swift-driver build for the Windows toolchain. As a
result, swift-collections fails to build properly when CMP0157 is set to
NEW due to object files not being generated.

This sets CMP0157 to OLD when targetting Android until the early
swift-driver is available on Windows.

This is similar to https://github.com/swiftlang/swift-corelibs-foundation/pull/5180