Change `Attachment.attach()` to `Attachment.record()` and `Attachable.withUnsafeBufferPointer()` to `Attachable.withUnsafeBytes()`.
This PR updates the `attach()` function to make it static on `Attachment` and
renames it `record()`, per the discussion [here](https://forums.swift.org/t/pitch-attachments/78072).

It also renames `Attachable.withUnsafeBufferPointer()` to `Attachable.withUnsafeBytes()`
to more closely match the convention used by the standard library for raw
buffers.

Finally, it removes the conditional dependence on the experimental language
feature "SuppressedAssociatedTypes" in `AttachableContainer`.