Capturing values in exit tests
This PR implements an experimental form of state capture in exit tests. If you
specify a capture list on the test's body closure and explicitly write the type
of each captured value, _and_ each value conforms to `Sendable`, `Codable`, and
(implicitly) `Copyable`, we'll encode them and send them "over the wire" to the
child process:

```swift
let a = Int.random(in: 100 ..< 200)
await #expect(exitsWith: .failure) { [a = a as Int] in
  assert(a > 500)
}
```

This PR is incomplete. Among other details:

- [] Need to properly transmit the data, not stuff it in an environment variable
- [] Need to capture source location information correctly for error handling
     during value decoding
- [] Need to implement diagnostics correctly

We are ultimately constrained by the language here as we don't have real type
information for the captured values, nor can we infer captures by inspecting the
syntax of the exit test body (hence the need for an explicit capture list with
types.) If we had something like `decltype()` we could apply during macro
expansion, you wouldn't need to write `x = x as T` and could just write `x`.