Add ConditionTrait.evaluate() (#909)
Add `ConditionTrait.evaluate()` so that a condition can be evaluated
independent of a `Test` object.

### Motivation:

Currently, the only way a `ConditionTrait` is evaluated is inside the
`prepare(for:)` method. This makes it difficult and awkward for
third-party libraries to utilize these traits because evaluating a
condition would require creating a dummy `Test` to pass to that method.

### Modifications:

Add `ConditionTrait.evaluate()`, and `ConditionTrait.Evaluation` enum
for the return value.

### Result:

Public API allows for evaluating a `ConditionTrait` in any context.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.