Pause child processes on spawn. (#1042)
This PR pauses child processes we spawn before resuming them, which
makes it easier to set a breakpoint and attach a debugger to those
processes. This functionality is available on macOS and Windows; as far
as I know, neither Linux nor POSIX-in-general has API for this.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.