Replace CRC-32 with SHA-256 in our macro target (#1047)
This PR replaces CRC-32 with SHA-256 in our macro target.

We currently use CRC-32 to disambiguate test functions that would
otherwise generate identical derived symbol names (using
`context.makeUniqueName()`.) By replacing it with the first 64 bits of a
SHA-256 hash, we reduce the odds of a collision.

We also currently use a 128-bit random number as a unique ID for exit
tests during macro expansion (which is as unique as it gets, but
unstable.) Replacing this random number with half of a SHA-256 hash
allows us to generate _stable_ IDs (that are still statistically unique)
which can help when debugging an exit test and may also improve cache
quality for tools (e.g. an IDE's symbol cache.)

Because there is no SHA-256 implementation available in the Swift
standard library or other components we can reliably link to in the
macro target, I've borrowed the implementation of SHA-256 in
swift-tools-support-core. The original version is
[here](https://github.com/swiftlang/swift-tools-support-core/blob/main/Sources/TSCBasic/HashAlgorithms.swift).

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.