Add signal handler for `SIGABRT_COMPAT` on Windows. (#1056)
The special-case signal handling implemented for Windows doesn't
currently set a handler for `SIGABRT_COMPAT` (which is a synonym of
`SIGABRT` with a different value.) This PR adds it to the list of
signals we install handlers for.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.