Don't use a class to store the current exit test.
This PR replaces the private `_CurrentContainer` class used to store the current
exit test with a bare pointer. The class is prone to duplicate definitions, but
we really just use it as a glorified box type for the move-only `ExitTest`, so
an immortal pointer will work just as well.

Works around rdar://148837303.