Don't use a class to store the current exit test. (#1065)
This PR replaces the private `_CurrentContainer` class used to store the
current exit test with a bare pointer. The class is prone to duplicate
definitions, but we really just use it as a glorified box type for the
move-only `ExitTest`, so an immortal pointer will work just as well.

Works around rdar://148837303.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.