Add `unsafe` keyword handling to macro expansions.
This PR changes the `@Test` and `#expect()` macros so they handle `unsafe`
expressions the same way `try` and `await` expressions are handled. The
propagation rules for `unsafe` aren't the same as for `try` and `await` (i.e. it
doesn't colour the calling function), but the general way we handle the keyword
is the same.

I haven't attempted to avoid inserting `unsafe` if a function is not marked
`@unsafe` as it complicates the necessary logic but has no effects at runtime.