Add `CustomStringConvertible` conformance to `ExitTest.Condition` and `StatusAtExit`.
This PR adds `CustomStringConvertible` conformance to these data/value types so
that they are presented reasonably in expectation failure messages or when
printed.

Strictly speaking, API changes should go through a review, but there's not
really anything to review here, so with the consent of the Testing Workgroup we
can amend the exit tests proposal to retroactively include these conformances.