Add `unsafe` keyword handling to macro expansions. (#1069)
This PR changes the `@Test` and `#expect()` macros so they handle
`unsafe` expressions the same way `try` and `await` expressions are
handled. The propagation rules for `unsafe` aren't the same as for `try`
and `await` (i.e. it doesn't colour the calling function), but the
general way we handle the keyword is the same.

I haven't attempted to avoid inserting `unsafe` if a function is not
marked `@unsafe` as it complicates the necessary logic but has no
effects at runtime.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.