[Experimental] Capturing values in exit tests (#1040)
This PR implements an experimental form of state capture in exit tests.
If you specify a capture list on the test's body closure and explicitly
write the type of each captured value, _and_ each value conforms to
`Sendable`, `Codable`, and (implicitly) `Copyable`, we'll encode them
and send them "over the wire" to the child process:

```swift
let a = Int.random(in: 100 ..< 200)
await #expect(exitsWith: .failure) { [a = a as Int] in
  assert(a > 500)
}
```

This PR is incomplete. Among other details:

- [x] Need to properly transmit the data, not stuff it in an environment
variable
- [x] Need to implement diagnostics correctly
- [x] Need to figure out if `ExitTest.CapturedValue` and
`__Expression.Value` have any synergy. _(They do, but it's beyond the
scope of this initial/experimental PR.)_

We are ultimately constrained by the language here as we don't have real
type information for the captured values, nor can we infer captures by
inspecting the syntax of the exit test body (hence the need for an
explicit capture list with types.) If we had something like `decltype()`
we could apply during macro expansion, you wouldn't need to write `x = x
as T` and could just write `x`. The macro would use `decltype()` to
produce a thunk function of the form:

```swift
@Sendable func __compiler_generated_name__(aʹ: decltype(a), bʹ: decltype(b), cʹ: decltype(c)) async throws {
  let (a, b, c) = (aʹ, bʹ, cʹ)
  try await { /* ... */ }()
}
```

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.