Enable experimental 'AllowUnsafeAttribute' feature to continue supporting 6.1 development snapshot toolchains (#1084)
This fixes a build failure when attempting to build the `main` branch
using a 6.1 development snapshot toolchain. This failure was introduced
by #1069, which added usage of the new `@unsafe` attribute, and the
failure was revealed when we set up the 6.1 CI jobs in #1083.

Here are some relevant related Swift PRs which give context around these
changes:

- https://github.com/swiftlang/swift/pull/75413
- https://github.com/swiftlang/swift/pull/79645

See the code comment for more details.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.