Fix build failure when building with pre-6.2 toolchain due to unrecognized `unsafe` keyword (#1093)
This fixes another bit of fallout from #1069 when building this
project's test targets with a 6.1 (or any pre-6.2) toolchain.

The `unsafe` keyword was introduced in 6.2 as part of [SE-0458: Opt-in
Strict Memory Safety
Checking](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0458-strict-memory-safety.md).
Older toolchains are not aware of it, so the fix is to avoid emitting
expressions involving that keyword when the macro plugin has been built
using an older toolchain.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.