Add withKnownIssue comments to known Issues (#1014)
Add withKnownIssue comments to known Issues

### Motivation:

The Comment passed to withKnownIssue() is currently only used when a
known issue does not occur. When a known issue does occur, it is marked
isKnown but does not contain any record of the withKnownIssue() comment,
making it harder to understand why the issue was considered to be known,
especially if there are multiple nested withKnownIssue() calls.

### Modifications:

When an issue is marked "known" by a `withKnownIssue()` call, the
recorded issue will now have a new `knownIssueContext` property
containing the comment passed to `withKnownIssue()` (if any). This
comment will be included in the `messages` array of the
`ABI.EncodedEvent` that represents the issue.

If the issue is recorded within multiple nested `withKnownIssue()`
calls, `knownIssueContext` corresponds to the innermost matching call.

The `Issue.isKnown` setter is now deprecated and a no-op.

When an error is thrown within a `withKnownIssue()` call, the `Issue`
passed to the issue matcher used to have the known issue comment in
`Issue.comments`. That has now been removed; Issues for thrown errors
now have no comments at all when passed to the issue matcher. This
matches Issues for thrown errors that occur outside of
`withKnownIssue()` calls.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.

---------

Co-authored-by: Jonathan Grynspan <grynspan@me.com>
Co-authored-by: Stuart Montgomery <smontgomery@apple.com>